requirements.txt
# Accent-Robust ASR — Requirements
# Python >= 3.9 required
# Install: pip install -r requirements.txt

# Core deep learning
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0

# HuggingFace ecosystem
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
tokenizers>=0.15.0
huggingface-hub>=0.20.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
audioread>=3.0.0
resampy>=0.4.2

# Speech evaluation
jiwer>=3.0.3           # WER/CER computation

# Data & ML utilities
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
Pillow>=10.0.0

# Config & utilities
PyYAML>=6.0.1
tqdm>=4.66.0
omegaconf>=2.3.0

# Notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.27.0

# Gradient checkpointing / efficient training
einops>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
